+++ tests/phpunit/Unit/OutputTest.php
@@ -8,7 +8,7 @@ use SFS\Output;
  * @covers  \SFS\Output
  * @group   semantic-forms-select
  *
- * @license GNU GPL v2+
+ * @license GPL-2.0-or-later
  * @since   1.3
  *
  * @author  mwjames